backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1